[ADM] Fix undo/redo selection behavior for widgets added from the palette
Also, fix a glitch in selection handling during a moveNode that resulted
from the selection changing once on the node's removal and again on the
node's reinsertion.

At the same time, fix a violation of the ADM/ADMNode abstraction barrier
so that ADMNode is not directly managing the selection in the ADM.

Finally, add a clearEvent call for selectionChanged to throw out any
obsolete events in the queue.